# swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Twitter Project API
  version: 1.0.0
...
paths:
  /tweets:
    post:
      tags:
        - tweets
      description: Запросом на этот endpoint пользователь будет создавать новый твит.Бэкенд будет его валидировать и сохранять в базу.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                api-key:
                  type: string
                tweet_data:
                  type: string
                tweet_media_ids:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: Successful creation
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                  tweet_id:
                    type: integer
        '500':
          description: Server error
    get:
      tags:
        - tweets
      description: Get tweet feed
      parameters:
        - name: api-key
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful retrieval
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                  tweets:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        content:
                          type: string
                        attachments:
                          type: array
                          items:
                            type: string
                        author:
                          type: object
                          properties:
                            id:
                              type: integer
                            name:
                              type: string
                        likes:
                          type: array
                          items:
                            type: object
                            properties:
                              user_id:
                                type: integer
                              name:
                                type: string
...